DjangoDemo/mysite/requirements.txt
aiohttp==2.2.5
aiomysql==0.0.9
alembic==0.9.6
async-timeout==1.4.0
beautifulsoup4==4.6.0
blinker==1.4
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
colorama==0.3.9
Django==1.11.7
django-haystack==2.6.1
Flask==0.12.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.6
itsdangerous==0.24
jieba==0.39
Jinja2==2.9.6
lxml==4.1.0
Mako==1.0.7
Markdown==2.6.10
MarkupSafe==1.0
multidict==3.2.0
mysql-connector==2.1.4
mysql-connector-python==2.0.4
olefile==0.44
Pillow==4.2.1
powerline-status==2.6
prettytable==0.7.2
pygame==1.9.3
Pygments==2.2.0
PyMySQL==0.7.11
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
reportlab==3.4.0
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.14
urllib3==1.22
virtualenv==15.1.0
Werkzeug==0.12.2
Whoosh==2.7.4
xlrd==1.1.0
xlwt==1.3.0
yarl==0.13.0
